Add options routes and rename route files
