+++ server/server.js
@@ -11,10 +11,11 @@ const Chat = require("./db/schema/03-chats");
 const Option = require("./db/schema/04-options");
 
 //import routes
-const chatsRoutes = require("./routes/chats")
-const logRoutes = require("./routes/log")
-const proposalsRoutes = require("./routes/proposals")
-const usersRoutes = require("./routes/users")
+const chatsRoutes = require("./routes/chatsRoutes")
+const logRoutes = require("./routes/logRoutes")
+const optionsRoutes = require("./routes/optionsRoutes")
+const proposalsRoutes = require("./routes/proposalsRoutes")
+const usersRoutes = require("./routes/usersRoutes")
 
 //Connect to database - the access string is imported from .env
 mongoose
@@ -32,6 +33,7 @@ app.use(express.json()) //Same purpose as body parser, lets server accept JSON a
 //-----Redirect to routes and pass them things imported above
 app.use("/chats", chatsRoutes(User, Chat, Proposal, bcrypt))
 app.use("/log", logRoutes(User, Chat, Proposal, bcrypt))
+app.use("/options", optionsRoutes(Option))
 app.use("/proposals", proposalsRoutes(User, Chat, Proposal, bcrypt))
 app.use("/users", usersRoutes(User, Chat, Proposal, bcrypt))
